Refactor to miminize rvi/mod.rs
This breaks out the service edge and the service handler into their own
submodules. The externally used objects and functions are publicly
reexported to present a coherent interface.
